rework agg primitives
