+++ featuretools/primitives/standard/__init__.py
@@ -1,4 +1,5 @@
 # flake8: noqa
+from featuretools.primitives.base.aggregation_primitive_base import AggregationPrimitive
 from featuretools.primitives.standard.aggregation_primitives import *
 from featuretools.primitives.standard.binary_transform import *
 from featuretools.primitives.standard.cum_transform_feature import *
